[Enhancement]: Extracted search feature into separate file.
