+++ firestore.rules
@@ -3,7 +3,7 @@ service cloud.firestore {
   match /databases/{database}/documents {
     match /restaurants/{name} {
       allow read, write: if
-          request.time < timestamp.date(2022, 6, 17);
+          request.time < timestamp.date(2023, 6, 17);
     }
   }
 }
